source refactoring
