+++ src/background.ts
@@ -5,26 +5,19 @@ import {
   protocol,
   BrowserWindow,
   ipcMain,
-  IpcMainEvent,
   MenuItem,
   Menu
 } from 'electron'
 import { createProtocol } from 'vue-cli-plugin-electron-builder/lib'
 import installExtension, { VUEJS_DEVTOOLS } from 'electron-devtools-installer'
 import { join } from 'path'
-import { LdapServer } from './library/LdapServer'
-import { LdapFactory } from './library/LdapFactory'
-import { LdapTree } from './library/LdapTree'
-import { SearchResult, Entry, Attribute } from 'ldapts'
-import { TreeNode, LdapConfig, LdapChange, ModifyAttributeTreeNodeObject, ModifyDnNodeObject, ModifyDnObject, DeleteDnObject, AddDnNodeObject, AddDnObject } from './library/common'
-import { Node } from 'tree-model'
-import { getAddAttributeList, getAttributeChanges, getDeleteDn, getModifyDn, getParentDn } from './library/LdapUtil'
+import { LdapServer } from './library/ldap/LdapServer'
+import { LdapFactory } from './library/ldap/LdapFactory'
+import { getAttributeTree, refreshRootTree, saveAllToLdap, serverBind, showAllChange } from './library/electron/IpcMainListener'
 
 const isDevelopment = process.env.NODE_ENV !== 'production'
 
 const mainIcon = join(__dirname, './assets/icons/icon.png')
-
-let isLdapConnected: boolean = false
 
 // Scheme must be registered before the app is ready
 protocol.registerSchemesAsPrivileged([
@@ -118,8 +111,9 @@ app.on('window-all-closed', async () => {
     app.quit()
   }
 
-  if (isLdapConnected) {
-    let ldapServer: LdapServer = LdapFactory.Instance()
+  let ldapServer: LdapServer = LdapFactory.Instance()
+
+  if (ldapServer.isConnected()) {  
     await ldapServer.disconnect()
   }
 })
@@ -147,159 +141,19 @@ app.on('ready', async () => {
 })
 
 // Ldap Connect and Get ldap entries
-ipcMain.on('serverBind', async (event : IpcMainEvent , ldapConfig : LdapConfig) => {
-  const ldapServer: LdapServer = LdapFactory.Instance()
-
-  if (isLdapConnected && ldapConfig !== ldapServer.ldapConfig) {
-    await ldapServer.disconnect()
-  }
-
-  ldapServer.ldapConfig = ldapConfig
-  const isAuthenticated: boolean = await ldapServer.connect()
-
-  if (isAuthenticated){
-    let searchResult: SearchResult | null = await ldapServer.search()
-    let ldapTree: LdapTree = new LdapTree()
-    if (searchResult){
-      ldapTree.makeEntryTree(ldapServer.baseDn, searchResult)
-      let rootNode : Node<TreeNode> | undefined = ldapTree.rootNode
-      if (rootNode) {
-        const searchResponse: Node<TreeNode>[] = [ rootNode.model ]
-        event.reply('allSearchResponse', searchResponse )
-      }
-    }
-  }
-
-  if (!isLdapConnected) {
-    isLdapConnected = true
-  }
-})
+ipcMain.addListener('serverBind', serverBind)
 
 // Get Attribute of ldap entry
-ipcMain.on("attributeTree", (event : IpcMainEvent, nodeName: string, nodeParent: TreeNode, isAddDn: Boolean, attributes?: Entry) => {
-  let ldapTree: LdapTree = new LdapTree()
-  let parentDn: string = getParentDn(nodeParent)
-  const attrResponse: TreeNode[] = ldapTree.makeAttrTree(nodeName, parentDn, attributes)
-  event.reply("attributeTreeResponse", attrResponse, isAddDn);
-})
+ipcMain.addListener('getAttributeTree', getAttributeTree )
 
-// Refresh ldap Entries
-ipcMain.on("refreshRootTree", async (event : IpcMainEvent) => {
-  const ldapServer: LdapServer = LdapFactory.Instance()
-  let searchResult: SearchResult | null = await ldapServer.search()
-  let ldapTree: LdapTree = new LdapTree()
-  if (searchResult){
-    ldapTree.makeEntryTree(ldapServer.baseDn, searchResult)
-    let rootNode : Node<TreeNode> | undefined = ldapTree.rootNode
-      if (rootNode) {
-        const searchResponse: Node<TreeNode>[] = [ rootNode.model ]
-        event.reply('allSearchResponse', searchResponse )
-      }
-  }
-})
+// Refresh All ldap Entries
+ipcMain.on("refreshRootTree", refreshRootTree)
 
-// Save All Changed Data
-ipcMain.on("saveAllChange", async (event: IpcMainEvent, addDnNodeList: AddDnNodeObject[], modifyDnNodeList: ModifyDnNodeObject[], saveAttributeList: ModifyAttributeTreeNodeObject[], deletDnNodeList: TreeNode[]) => {
-  const ldapServer: LdapServer = LdapFactory.Instance()
+// Save All Changed Data To Ldap
+ipcMain.on("saveAllToLdap", saveAllToLdap)
 
-  if (addDnNodeList.length > 0) {
-    addDnNodeList.forEach(async (addDnNodeObject: AddDnNodeObject) => {
-      let [dn, attrList] = getAddAttributeList(addDnNodeObject.nodeName, addDnNodeObject.attrTree)
-      if (ldapServer.isConnected()) {
-        await ldapServer.add(dn, attrList)
-      }
-    })
-  }
-
-  if (modifyDnNodeList.length > 0) {
-    modifyDnNodeList.forEach(async (modifyDnNodeObject: ModifyDnNodeObject) => {
-      let [nodeDn, modifyDn] = getModifyDn(modifyDnNodeObject)
-      if (ldapServer.isConnected()) {
-        await ldapServer.modifyDn(nodeDn, modifyDn)
-      }
-    })
-  }
-
-  if (saveAttributeList.length > 0) {
-    saveAttributeList.forEach(async (attribute: ModifyAttributeTreeNodeObject) => {
-      let attrTree: TreeNode[] = attribute.tree
-      let deleteList: TreeNode[] = attribute.deleteList
-      let changeData: LdapChange = getAttributeChanges(attrTree, deleteList)
-
-      if (ldapServer.isConnected()) {
-        await ldapServer.modify(changeData);
-      }
-    })
-  }
-
-  if (deletDnNodeList.length > 0) {
-    deletDnNodeList.forEach(async (deleteDnNode: TreeNode) => {
-      let [originDn, parentDn] = getDeleteDn(deleteDnNode)
-
-      if (ldapServer.isConnected()) {
-        await ldapServer.delete(originDn)
-      }
-    })
-  }
-
-  await new Promise(resolve => setTimeout(resolve, 1000));
-  event.reply("refreshRootTreeFromMain");
-})
-
-ipcMain.on("showChangePage", async (event : IpcMainEvent, addDnNodeList: AddDnNodeObject[], modifyDnNodeList : ModifyDnNodeObject[], saveAttributeList: ModifyAttributeTreeNodeObject[], deleteDnNodeList: TreeNode[]) => {
-  const ldapServer: LdapServer = LdapFactory.Instance()
-
-  let addDnList: AddDnObject[] = []
-  let modifyDnList: ModifyDnObject[] = []
-  let changeAttrList: LdapChange[] = []
-  let deleteDnList: DeleteDnObject[] = []
-
-  if (addDnNodeList.length > 0) {
-    addDnNodeList.forEach((addDnNode: AddDnNodeObject) => {
-      let [dn, attrList]: [string, Attribute[]] = getAddAttributeList(addDnNode.nodeName, addDnNode.attrTree)
-
-      addDnList.push({
-        dn: dn,
-        attrList: attrList
-      })
-    })
-  }
-
-  if (modifyDnNodeList.length > 0) {
-    modifyDnNodeList.forEach((modifyDnNodeObject: ModifyDnNodeObject) => {
-      let [originDn, modifyDn]: [string, string] = getModifyDn(modifyDnNodeObject)
-
-      modifyDnList.push({
-        originDn: originDn,
-        modifyDn: modifyDn
-      })
-    })
-  }
-
-  if (saveAttributeList.length > 0) {
-    saveAttributeList.forEach((attribute: ModifyAttributeTreeNodeObject) => {
-      let attrTree: TreeNode[] = attribute.tree
-      let deleteList: TreeNode[] = attribute.deleteList
-      let changeData: LdapChange = getAttributeChanges(attrTree, deleteList)
-      changeAttrList.push(changeData)
-    })
-  }
-
-  if (deleteDnNodeList.length > 0) {
-    deleteDnNodeList.forEach((deleteDnNode: TreeNode) => {
-      let [originDn, parentDn]: [string, string] = getDeleteDn(deleteDnNode)
-
-      if (originDn !== '' && parentDn !== '') {
-        deleteDnList.push({
-          originDn: originDn,
-          parentDn: parentDn
-        })
-      }
-    })
-  }
-  
-  event.reply("returnShowChangePage", addDnList, modifyDnList, changeAttrList, deleteDnList);
-})
+// Show All Change
+ipcMain.on("showAllChange", showAllChange )
 
 // Exit cleanly on request from parent process in development mode.
 if (isDevelopment) {
